feat: Auth / Pass loginHandler props from Auth #147
